structure refactoring
