+++ boolean.go
@@ -1,0 +1,31 @@
+// Copyright (c) 2016 ~ 2019, Alex Stocks.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package hessian
+
+/////////////////////////////////////////
+// Bool
+/////////////////////////////////////////
+
+// # boolean true/false
+// ::= 'T'
+// ::= 'F'
+func encBool(b []byte, v bool) []byte {
+	var c byte = BC_FALSE
+	if v == true {
+		c = BC_TRUE
+	}
+
+	return append(b, c)
+}
