+++ encode_test.go
@@ -1,0 +1,40 @@
+// Copyright (c) 2016 ~ 2019, Alex Stocks.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package hessian
+
+import (
+	"bytes"
+	"testing"
+)
+
+// go test -v encode_test.go encode.go decode.go const.go codec.go pojo.go
+var assertEqual = func(want, got []byte, t *testing.T) {
+	if !bytes.Equal(want, got) {
+		t.Fatalf("want %v , got %v", want, got)
+	}
+}
+
+func TestEncNull(t *testing.T) {
+	var (
+		e *Encoder
+	)
+
+	e = NewEncoder()
+	e.Encode(nil)
+	if e.Buffer() == nil {
+		t.Fail()
+	}
+	t.Logf("nil enc result:%s\n", string(e.buffer))
+}
